==> 5.1_firm_agg.ipynb <==
# >>> Path configuration (auto-inserted) >>>
from pathlib import Path
import os

PROJECT_ROOT = Path(os.getenv("GBB_PROJECT_ROOT", ".")).resolve()
DATA = PROJECT_ROOT / "data"
PATSTAT = PROJECT_ROOT / "patstat"
PATTEXT = PROJECT_ROOT / "patent_text"
SAMPLEDATA = PROJECT_ROOT / "sampledata"

# Fallback to sampledata if primary paths not present
if not PATSTAT.exists() and (SAMPLEDATA / "patstat").exists():
    PATSTAT = SAMPLEDATA / "patstat"
if not PATTEXT.exists() and (SAMPLEDATA / "patent_text").exists():
    PATTEXT = SAMPLEDATA / "patent_text"
if not DATA.exists() and (PROJECT_ROOT / "data").exists():
    DATA = PROJECT_ROOT / "data"

def P(*parts):
    return str(Path(*parts))
# <<< Path configuration (auto-inserted) <<<


import pandas as pd
import numpy as np

firmidname = pd.read_parquet(
    str(PATSTAT / "tls206.parquet"),
    columns=["person_id", "psn_id", "psn_name", "psn_sector"],
).query("psn_sector=='COMPANY'").drop(columns='psn_sector')
firmidname.head()

firmidname.to_parquet('firmidname.parquet',index=False)

tls207 = pd.read_parquet(
    str(PATSTAT / "tls207.parquet"),
    columns=["person_id", "appln_id"],
)
tls207.head()

tls201 = pd.read_parquet(
    str(PATSTAT / "tls201.parquet"),
    columns=["appln_id","docdb_family_id",'appln_filing_year']
)
tls201.head()

firmpat = (
    tls201.merge(tls207)
    .merge(firmidname[["person_id", "psn_id"]].drop_duplicates())
    .groupby(["psn_id", "docdb_family_id"])["appln_filing_year"]
    .min()
    .reset_index()
    .rename(columns={"psn_id": "person_id", "appln_filing_year": "year"})
)
firmpat.head()

firmpat["period"] = np.where(firmpat.year.between(1985, 1994), 1985, 0)
firmpat["period"] = np.where(firmpat.year.between(1995, 2004), 1995, firmpat.period)
firmpat["period"] = np.where(firmpat.year.between(2005, 2014), 2005, firmpat.period)
firmpat["period"] = np.where(firmpat.year >= 2015, 0, firmpat.period)
firmpat = firmpat.merge(
    firmpat.groupby("person_id")["docdb_family_id"]
    .nunique()
    .reset_index()
    .rename(columns={"docdb_family_id": "totalpat"})
)
firmpat.head()

firmpat = firmpat.query('totalpat>=100')
firmpat.shape, firmpat.person_id.nunique()

firmpat.to_parquet('firmpat.parquet')

firmpat = pd.read_parquet('firmpat.parquet')
firmpat.head()

firmpat.query('totalpat>=100').person_id.nunique()

firmpat = firmpat.query('totalpat>=100')
firmpat.person_id.nunique()

firmpat.to_parquet('firmpat2.parquet',index=False)

## start here

firmpat = pd.read_parquet('firmpat2.parquet')
firmpat.head()

fpc = firmpat.groupby(['person_id','period'])['docdb_family_id'].nunique().reset_index().rename(columns={'docdb_family_id':'wt'})
fpc.head()

fpc.to_parquet('firm_period_patcount.parquet')

tls225 = pd.read_parquet(str(PATSTAT / "tls225.parquet"),columns=['docdb_family_id','cpc_class_symbol','cpc4d'])
tls225.head()

firm_cpc = firmpat[['person_id','docdb_family_id','period']].drop_duplicates().merge(tls225).rename(columns={'cpc_class_symbol':'cpc'})
firm_cpc.head()

firm_cpc.to_parquet('firm_pat_cpc.parquet')

firmperiody02 = (
    firm_cpc.assign(y02=firm_cpc.cpc.str.split('/').str.get(0))[firm_cpc.cpc.str[:3].isin(["Y02",'Y04'])]
    .groupby(["person_id", "period", "y02"])["docdb_family_id"]
    .nunique()
    .reset_index()
    .rename(columns={"docdb_family_id": "num_pat_ccmt"})
)
firmperiody02.head()

firmperiody02.to_parquet('firmperiody02.parquet')

cluscpc = pd.read_parquet(str(DATA /'clusangle_outlier_hdbscan2.parquet')).query('clus!="c-1"')
#cluscpc['cpc'] = cluscpc['cpc'].str.replace(' ','')
#cluscpc['num'] = cluscpc.groupby('clus')['cpc'].transform('nunique')
#cluscpc = cluscpc.assign(clus="c" + cluscpc.clus.astype(str))
cluscpc.head()

cluscpc['rnk'] = cluscpc.groupby('clus')['max'].rank(ascending=False)
cluscpc['rca'] = cluscpc['max']*cluscpc['max'].sum()/cluscpc.groupby('clus')['max'].transform(sum)/cluscpc.groupby('cpc')['max'].transform(sum)
##note the filtering!
#cluscpc = cluscpc.query('count>5')
cluscpc.head()

cluscpc.cpc.nunique()

srccpc = pd.read_parquet(str(DATA /'clusangle_outlier_hdbscan2_src2.parquet')).query('clus!="s-1"')
srccpc.head()

firmperiodcpc = firm_cpc.drop_duplicates(subset=["person_id", "period", "docdb_family_id",'cpc']).groupby(['person_id','period','cpc'])['docdb_family_id'].nunique().reset_index()
firmperiodcpc['cpc'] = firmperiodcpc['cpc'].str.replace(' ','')
firmperiodcpc.head()

firmperiodcpc.to_parquet('firm_cpc_period.parquet')

firmperiodclus = firmperiodcpc.merge(cluscpc[['clus','cpc']])
firmperiodclus.head()

firmperiodclusrca = firmperiodclus.groupby(['person_id','period','clus'])['docdb_family_id'].sum().reset_index()
firmperiodclusrca['rca'] = (
    firmperiodclusrca.groupby("period")["docdb_family_id"].transform(sum)
    * firmperiodclusrca.docdb_family_id
    / firmperiodclusrca.groupby(["person_id", "period"])["docdb_family_id"].transform(sum)
    / firmperiodclusrca.groupby(["clus", "period"])["docdb_family_id"].transform(sum)
)
firmperiodclusrca['binrca'] = np.where(firmperiodclusrca.rca>=1,1,0)
firmperiodclusrca.head()

firmperiodcluscnt = firmperiodclusrca.query('binrca>=1').groupby(['person_id','period'])['clus'].nunique().reset_index().rename(columns={"clus":'num_clus'})
firmperiodcluscnt.head()

firmperiodcluscnt.to_parquet('firmperiodclus.parquet')

firmperiodsrc = firmperiodcpc.merge(srccpc[['clus','cpc']])
firmperiodsrc.head()

firmperiodsrcrca = firmperiodsrc.rename(columns={'clus':'src'}).groupby(['person_id','period','src'])['docdb_family_id'].sum().reset_index()
firmperiodsrcrca['rca'] = (
    firmperiodsrcrca.groupby("period")["docdb_family_id"].transform(sum)
    * firmperiodsrcrca.docdb_family_id
    / firmperiodsrcrca.groupby(["person_id", "period"])["docdb_family_id"].transform(sum)
    / firmperiodsrcrca.groupby(["src", "period"])["docdb_family_id"].transform(sum)
)
firmperiodsrcrca['binrca'] = np.where(firmperiodsrcrca.rca>=1,1,0)
firmperiodsrcrca.head()

firmperiodsrccnt = firmperiodsrcrca.query('binrca>=1').groupby(['person_id','period'])['src'].nunique().reset_index().rename(columns={"src":'num_src'})
firmperiodsrccnt.head()

firmperiodsrccnt.to_parquet('firmperiodsrc.parquet')

cpc4dclus = pd.read_parquet(str(DATA /'srcpath33.parquet')).query('path!="c-1" and src!="s-1"')
#cpc4dclus['clus'] = 'c'+cpc4dclus['clus'].astype(str)
cpc4dclus.head()

cpc4dclus2 = cpc4dclus[(cpc4dclus.binrca==1)]
#cpc4dclus2 = cpc4dclus[(cpc4dclus.rnk<cpc4dclus.effhhi)]
cpc4dclus2.path.nunique(),cpc4dclus2.src.nunique()

firmperiodsrcpath = (
    firmperiodsrcrca[firmperiodsrcrca.binrca == 1][["person_id", "src", "period"]]
    .merge(cpc4dclus2[["src", "path"]].drop_duplicates())
    .merge(firmperiodclusrca.rename(columns={"clus":'path'}).query('binrca>=1')[["person_id", "path", "period"]].drop_duplicates().assign(exist=1),how='left')
)
firmperiodsrcpath['exist'] = firmperiodsrcpath['exist'].fillna(0)
firmperiodsrcpath.head()

firmperiodsrcpath.to_parquet('firmperiodsrcpath.parquet')

clus2y02 = pd.read_parquet(str(DATA /'y02path33.parquet')).query('path!="c-1"')
#clus2y02['clus'] = 'c'+clus2y02['clus'].astype(str)
clus2y02 = clus2y02[clus2y02.binrca==1]
clus2y02.head()

firmperiody02clus = (
    firmperiodclusrca.query("binrca>=1")[["person_id", "clus", "period"]].rename(columns={'clus':'path'})
    .drop_duplicates()
    .merge(clus2y02[["path", "y02"]].drop_duplicates())
)
firmperiody02clus.head()

firmperiody02clus.to_parquet('firmperiody02path.parquet',index=False)

firmperiody02cluscnt = firmperiody02clus.groupby(['person_id','period','y02'])['path'].nunique().reset_index()
firmperiody02cluscnt.head()

firmperiody02cluscnt.to_parquet('firmperiody02pathcnt.parquet',index=False)

firmperiody02srcpath = firmperiody02clus.merge(
    firmperiodsrcrca[firmperiodsrcrca.binrca == 1][["person_id", "src", "period"]]
    .merge(cpc4dclus2[["src", "path"]].drop_duplicates())
    .merge(
        firmperiodclusrca.query("binrca>=1")[["person_id", "clus", "period"]]
        .drop_duplicates().rename(columns={'clus':'path'})
    )
)
firmperiody02srcpath.head()

firmperiody02srcpath.to_parquet('firmperiody02srcpath.parquet')

firmperiody02srcpathcnt = firmperiody02srcpath.groupby(['person_id','period','y02'])['src'].nunique().reset_index()
firmperiody02srcpathcnt.head()

firmperiody02srcpathcnt.to_parquet('firmperiody02srcpathcnt.parquet')

clus2y02.head()

firmperiodsrcrca.head()

firmperiody02src = clus2y02.query('binrca==1')[['path','y02']].drop_duplicates().merge(
    firmperiodsrcrca[firmperiodsrcrca.binrca == 1][["person_id", "src", "period"]]
    .merge(cpc4dclus2[["src", "path"]].drop_duplicates())
)
firmperiody02src.head()

firmperiody02clussrc = firmperiody02clus.merge(
    firmperiodsrcrca[firmperiodsrcrca.binrca == 1][["person_id", "src", "period"]]
    .merge(cpc4dclus2[["src", "path"]].drop_duplicates())
    .merge(
        firmperiodclusrca.query("binrca>=1")[["person_id", "clus", "period"]]
        .drop_duplicates().rename(columns={'clus':'path'})
    )
)
firmperiody02clussrc.head()

firmperiody02srccnt = firmperiody02src.groupby(['person_id','period','y02'])['src'].nunique().reset_index()
firmperiody02srccnt.head()

firmperiody02srccnt.to_parquet('firmperiody02srccnt.parquet')
