==> 5.3_entry_reg_R.ipynb <==
# >>> Path configuration (auto-inserted) >>>
PROJECT_ROOT <- normalizePath(Sys.getenv("GBB_PROJECT_ROOT", "."), mustWork = FALSE)
DATA <- file.path(PROJECT_ROOT, "data")
PATSTAT <- file.path(PROJECT_ROOT, "patstat")
PATTEXT <- file.path(PROJECT_ROOT, "patent_text")
SAMPLEDATA <- file.path(PROJECT_ROOT, "sampledata")

if (!dir.exists(PATSTAT) && dir.exists(file.path(SAMPLEDATA, "patstat"))) {
  PATSTAT <- file.path(SAMPLEDATA, "patstat")
}
if (!dir.exists(PATTEXT) && dir.exists(file.path(SAMPLEDATA, "patent_text"))) {
  PATTEXT <- file.path(SAMPLEDATA, "patent_text")
}
if (!dir.exists(DATA) && dir.exists(file.path(PROJECT_ROOT, "data"))) {
  DATA <- file.path(PROJECT_ROOT, "data")
}
# <<< Path configuration (auto-inserted) <<<


## install below if not exist
## install.packages(c("tidyverse","fixest","arrow"))
library(tidyverse)
library(fixest)
library(arrow)

growthdf <- read_parquet('firm_green_data.parquet') |> filter(period==1995) |> filter(wt>=10) 
growthdf |> head()

mod1 = feols((num_pat_ccmt_2>0) ~ log(path+1)*log(src+1)|person_id+y02,cluster = ~person_id+y02,data= growthdf|>filter(num_pat_ccmt==0))
print(mod1)

mod2 = feols((num_pat_ccmt_2>0) ~ log(path+1)+log(src+1)|person_id+y02,cluster = ~person_id+y02,data= growthdf|>filter(num_pat_ccmt==0))
print(mod2)

mod3 = feols((num_pat_ccmt_2>0) ~ log(path+1)|person_id+y02,cluster = ~person_id+y02,data= growthdf|>filter(num_pat_ccmt==0))
print(mod3)

mod4 = feols((num_pat_ccmt_2>0) ~ log(src+1)|person_id+y02,cluster = ~person_id+y02,data= growthdf|>filter(num_pat_ccmt==0))
print(mod4)

etable(mod4,mod3,mod2,mod1)
