==> 5.4_marginplot_stata.ipynb <==
# Path note (manual)
# This notebook contains Stata cells. Define globals in Stata before running:
# global PROJECT_ROOT "."
# global DATA "$PROJECT_ROOT/data"
# global PATSTAT "$PROJECT_ROOT/patstat"
# global PATTEXT "$PROJECT_ROOT/patent_text"
# Optionally point them to "$PROJECT_ROOT/sampledata/..." for the demo.


use firm_green_entry.dta

gen depvar = (num_pat_ccmt_2>0)*1

gen nsrc = log(src+1)

gen npath = log(path+1)

sum npath,d
local L `r(p5)'
local H `r(p95)'
local STEP = (`H'-`L')/10
reghdfe depvar c.nsrc##c.npath, absorb (person_id y02) cluster(person_id y02) residuals nocons
margins, dydx(nsrc) at(npath=(`L'(`STEP')`H')) predict(xb)
marginsplot, recast(line) recastci(rarea) plotopts(lcolor(blue) lpattern(solid) lwidth(medium) fcolor(navy%30) ) name(margins_npath, replace) xlabel(#5) title("marginal effect of starting points") ytitle("{&part}log(#patents+1)/{&part}log(#starting points+1)") xtitle("log(#pathways+1)")  aspect(1) scheme(s1mono)
graph export "margins_nsrc.png", replace
graph export "margins_nsrc.pdf", replace

sum nsrc,d
local L `r(p5)'
local H `r(p95)'
local STEP = (`H'-`L')/10
reghdfe depvar c.nsrc##c.npath, absorb (person_id y02) cluster(person_id y02) residuals nocons
margins, dydx(npath) at(nsrc=(`L'(`STEP')`H')) predict(xb)
marginsplot, recast(line) recastci(rarea) plotopts(lcolor(blue) lpattern(solid) lwidth(medium) fcolor(navy%30)) name(margins_nsrc, replace)  xlabel(#5) title("marginal effect of pathways") ytitle("{&part}log(#patents+1)/{&part}log(#pathways+1)") xtitle("log(#starting points+1)") aspect(1) scheme(s1mono)
graph export "margins_npath.png", replace
graph export "margins_npath.pdf", replace

graph combine margins_npath margins_nsrc, col(2) scheme(s1mono) ycommon ysize(7.5cm) xsize(16cm) altshrink
graph export "margins_combined.png", replace
graph export "margins_combined.pdf", replace
